style: clean directory scripts
create scripts/Tests directory and shift all parameter testings of the constraints algorithm and the first parallelizaton tests into it. Only the python files are stored. Test results
for each example are from now on stored and analysed on the tubcloud (..\DBTA\MOSAIC\ModOpt\pyModOpt\Tests\). They will be collected in directories named as Result_<index>. Hence no
directory of this type will be synchronized on git which is stated in the .gitignore file.