add: parallelization method to partial interval nesting procedure
created new module iNes_procedure.py which is valid for complete and partial
procedure from now on. Everything needed for parallelization is in the new
module parallelization.py. iNes_procedure.py and parallelization.py are invoked
in main.py where the user settings are processed. Instead of the algorithm settings
"complete", "complete_seq" and "partial", only "complete" and "partial" remain. The
parallelization is set by the user through dict_options["parallelization"] as True/False.

style: documentation update in the new and modified files