style: split variable and branch parallelization
user can now independently choose if variables or branches are parallelized. On top the
number of cores used in the programm is also set by the user. These three handles are now
part of the UDLS in MOSAIC (ID: 115594).

fix: functions without addition

a function like: 0 = a*x can be handled now. It is slightly modified in iNes_procedure/splitFunctionByVariableDependency(...):
0 = a*x + eps/2 with eps being the machine accuracy. Adding this very small value the function can be properly splitted into an
x-dependent and x-independent part.