Fix: second derrivate for non-continuous section identification not suifficient
because df2dx > 0 and df2dx <0 is also valid for f(x)=x^3 for example. New idea is the identification by first (or second derrivative) if dfdx = [-inf, b] or [-inf, +inf] or [a,+inf]. Step functions and poles could be filtered out whenever this is the case in a first step. Secondly the already implemented monotone section identification is started.

TODO: Implementation of filtering method look for TODO in INes_procedure