Fix: Removed boundsAlmostEqual array to increase reduction efficiency Fix: Function extensions with true discontinuous variables intervals are not used for their reduction
First fix was considered since for sets of interval vectors if variable intervals in one branch don't
change anymore their entry in boundsAlmostEqual was set to True and this array was also used for the
reduction of the next inerval vectors. Hence, their variable intervals might not have been further reduced
even if they could have been still reduced.

Second fix was necessary to not interupt the reduction process for example for x**2/[0,1]. Its upper bound will always
be +inf no matter how the interval x looks like. Therefore this expression is not useful to reduce x at all. Nevertheless
instead of returning an empty set and causing a termination of the program, the complete interval x is returned, because in
the next iteration the singularity in 1/[0,1] might disapear through reducing the other variable intervals.